+++ packages/react/src/pagination/PaginationContainer.tsx
@@ -1,6 +1,21 @@
 import type { PropsWithChildren } from 'react';
 import { Flex } from '../flex';
 
+export type PaginationItemContainerProps = PropsWithChildren<{
+	'aria-label': string;
+}>;
+
+export const PaginationItemContainer = ({
+	children,
+	'aria-label': ariaLabel,
+}: PaginationItemContainerProps) => (
+	<nav aria-label={ariaLabel}>
+		<Flex as="ol" alignItems="center" justifyContent="center" flexWrap="wrap">
+			{children}
+		</Flex>
+	</nav>
+);
+
 export type PaginationContainerProps = PropsWithChildren<{
 	hasRightArea: boolean;
 }>;
